requirements.txt
weaviate-client
flask
flask-cors
openai